tune alloc_f16_buf
